Flytter dataklasser ut av route
